🚚 (api): move CenterHabilitationError to certification shared domain
