+++ api/src/certification/evaluation/domain/errors.js
@@ -12,11 +12,4 @@ class CertificationComputeError extends DomainError {
   }
 }
 
-class CenterHabilitationError extends DomainError {
-  constructor() {
-    super('This certification center has no habilitation for the given complementary certification.');
-    this.code = 'CENTER_HABILITATION_ERROR';
-  }
-}
-
-export { CenterHabilitationError, CertificationComputeError, ChallengeAlreadyAnsweredError };
+export { CertificationComputeError, ChallengeAlreadyAnsweredError };
