+++ api/src/certification/shared/application/http-error-mapper-configuration.js
@@ -2,13 +2,16 @@ import { HttpErrors } from '../../../shared/application/http-errors.js';
 import { DomainErrorMappingConfiguration } from '../../../shared/application/models/domain-error-mapping-configuration.js';
 import { configurationDomainErrorMappingConfiguration } from '../../configuration/application/http-error-mapper-configuration.js';
 import { enrolmentDomainErrorMappingConfiguration } from '../../enrolment/application/http-error-mapper-configuration.js';
-import { evaluationDomainErrorMappingConfiguration } from '../../evaluation/application/http-error-mapper-configuration.js';
 import {
   parcoursupDomainErrorMappingConfiguration,
   resultsDomainErrorMappingConfiguration,
 } from '../../results/application/http-error-mapper-configuration.js';
 import { sessionDomainErrorMappingConfiguration } from '../../session-management/application/http-error-mapper-configuration.js';
-import { CertificationCourseUpdateError, InvalidCertificationReportForFinalization } from '../domain/errors.js';
+import {
+  CenterHabilitationError,
+  CertificationCourseUpdateError,
+  InvalidCertificationReportForFinalization,
+} from '../domain/errors.js';
 
 const certificationDomainErrorMappingConfiguration = [
   {
@@ -21,6 +24,10 @@ const certificationDomainErrorMappingConfiguration = [
     name: CertificationCourseUpdateError.name,
     httpErrorFn: (error) => new HttpErrors.BadRequestError(error.message, error.code, error.meta),
   },
+  {
+    name: CenterHabilitationError.name,
+    httpErrorFn: (error) => new HttpErrors.ForbiddenError(error.message, error.code),
+  },
 ].map((domainErrorMappingConfiguration) => new DomainErrorMappingConfiguration(domainErrorMappingConfiguration));
 
 certificationDomainErrorMappingConfiguration.push(
@@ -29,6 +36,5 @@ certificationDomainErrorMappingConfiguration.push(
   ...enrolmentDomainErrorMappingConfiguration,
   ...sessionDomainErrorMappingConfiguration,
   ...configurationDomainErrorMappingConfiguration,
-  ...evaluationDomainErrorMappingConfiguration,
 );
 export { certificationDomainErrorMappingConfiguration };
